+++ src/components/sbb-checkbox-group/sbb-checkbox-group.custom.ts
@@ -1,0 +1,5 @@
+export interface InterfaceSbbCheckboxGroupAttributes {
+  horizontalFrom?: 'zero' | 'micro' | 'small' | 'medium' | 'large' | 'wide' | 'ultra';
+  orientation: 'horizontal' | 'vertical';
+  size: 'm' | 's';
+}