+++ src/components/sbb-toggle-check/sbb-toggle-check.custom.d.ts
@@ -1,3 +1,8 @@
 export interface InterfaceToggleCheckAttributes {
   labelPosition?: 'before' | 'after';
 }
+
+export interface SbbToggleCheckChange {
+  checked: boolean;
+  value: string;
+}